+++ src/org/elixir_lang/model/psi/protocol/ProtocolFunction.kt
@@ -13,7 +13,6 @@ import com.intellij.psi.PsiFile
 import com.intellij.psi.ResolveState
 import com.intellij.psi.SmartPointerManager
 import com.intellij.psi.search.SearchScope
-import com.intellij.refactoring.rename.api.RenameTarget
 import com.intellij.util.concurrency.annotations.RequiresReadLock
 import org.elixir_lang.model.psi.ElixirSymbolWithUsages
 import org.elixir_lang.psi.CallDefinitionClause
@@ -42,7 +41,7 @@ class ProtocolFunction(
     val name: String,
     val arity: Int,
     val macro: Boolean
-) : ElixirSymbolWithUsages, NavigationTarget, SearchTarget, RenameTarget {
+) : ElixirSymbolWithUsages, NavigationTarget, SearchTarget {
 
     override val searchText: String get() = name
     override val targetName: String get() = name